Revert "Working on setting up new repository connection"
This reverts commit e6ff78b19d155d2aaaae4a99c90299dc01203e4a.

+++ src/main/java/com/mellora/rpgbot/dao/GuildSettingsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.mellora.rpgbot.dao.models.GuildSettings;
 
-public interface GuildSettingsRepository extends CrudRepository<GuildSettings, Integer>{
+public interface GuildSettingsRepository extends CrudRepository<GuildSettings, Long>{
 
 }
